File reading test in a separated mod
