+++ src/dos_tests/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod file;
